Introduce CloudRequest
