+++ FeedHenry.podspec
@@ -1,6 +1,6 @@
 Pod::Spec.new do |s|
   s.name         = 'FeedHenry'
-  s.version      = '4.0.0-alpha'
+  s.version      = '4.0.0-alpha1'
   s.summary      = 'FeedHenry Swift iOS Software Development Kit'
   s.homepage     = 'https://www.feedhenry.com'
   s.social_media_url = 'https://twitter.com/feedhenry'
